Q1-Q5/D1-D3/A1-A3/C1-C3/P1-P4/O1-O3: Remaining items batch
Q1: Covered by B1 (delete_sprint ? operator).
Q2: Reviewed — escape_csv is only 10 lines, not worth extracting.
Q3: Remove redundant 'as i64' casts in epics.rs (4 occurrences).
Q4: Covered by B13 (auto_archive_days added to frontend Config).
Q5: Remove dead dependency handlers from watchers.rs — the actual
  handlers live in dependencies.rs with typed request structs.

D1: API changelog updated with v11 FTS5, backup/restore, user
  hours report, and other improvements.
D2: Reviewed — FTS5 now has LIKE fallback (B12), so no special
  SQLite build required.
D3: Register import_tasks_csv, export_burns, health in OpenAPI.

P1: Reviewed — voted_tids empty check already guards the query.
P2: Reviewed — team scope CTE acceptable for project scale.
P3: Reviewed — ETag subqueries are fast on indexed columns.
P4: Reviewed — audit endpoint already supports user_id filter.

O1: Reviewed — integrity_check is slow on large DBs, not worth
  running on every startup.
O2: Reviewed — FTS5 availability is now tracked (B12).
O3: Reviewed — log rotation is an ops concern (systemd/logrotate).

A1: Already implemented — SVG has role=progressbar with aria attrs.
A2: Reviewed — board cards have role=list on columns.
A3: Shortcuts modal now dismissible with Escape key.

C1: Remove unused _state parameter from list_backups.
C2: Remove dead token field from SseQuery.
C3: Reviewed — RoomMember.user_id is sent by backend.

236 backend tests, 154 frontend tests, TS strict clean.

+++ crates/pomodoro-daemon/src/routes/epics.rs
@@ -30,7 +30,7 @@ pub async fn get_epic_group(State(engine): State<AppState>, _claims: Claims, Pat
 #[utoipa::path(delete, path = "/api/epics/{id}", responses((status = 204)), security(("bearer" = [])))]
 pub async fn delete_epic_group(State(engine): State<AppState>, claims: Claims, Path(id): Path<i64>) -> Result<StatusCode, ApiError> {
     let detail = db::get_epic_group_detail(&engine.pool, id).await.map_err(|_| err(StatusCode::NOT_FOUND, "Epic group not found"))?;
-    if detail.group.created_by != claims.user_id as i64 && claims.role != "root" {
+    if detail.group.created_by != claims.user_id && claims.role != "root" {
         return Err(err(StatusCode::FORBIDDEN, "Not owner"));
     }
     db::delete_epic_group(&engine.pool, id).await.map_err(internal)?;
@@ -40,7 +40,7 @@ pub async fn delete_epic_group(State(engine): State<AppState>, claims: Claims, P
 #[utoipa::path(post, path = "/api/epics/{id}/tasks", responses((status = 204)), security(("bearer" = [])))]
 pub async fn add_epic_group_tasks(State(engine): State<AppState>, claims: Claims, Path(id): Path<i64>, Json(req): Json<EpicGroupTasksRequest>) -> Result<StatusCode, ApiError> {
     let detail = db::get_epic_group_detail(&engine.pool, id).await.map_err(|_| err(StatusCode::NOT_FOUND, "Epic group not found"))?;
-    if detail.group.created_by != claims.user_id as i64 && claims.role != "root" {
+    if detail.group.created_by != claims.user_id && claims.role != "root" {
         return Err(err(StatusCode::FORBIDDEN, "Not owner"));
     }
     if req.task_ids.is_empty() { return Ok(StatusCode::NO_CONTENT); }
@@ -60,7 +60,7 @@ pub async fn add_epic_group_tasks(State(engine): State<AppState>, claims: Claims
 #[utoipa::path(delete, path = "/api/epics/{id}/tasks/{task_id}", responses((status = 204)), security(("bearer" = [])))]
 pub async fn remove_epic_group_task(State(engine): State<AppState>, claims: Claims, Path((id, task_id)): Path<(i64, i64)>) -> Result<StatusCode, ApiError> {
     let detail = db::get_epic_group_detail(&engine.pool, id).await.map_err(|_| err(StatusCode::NOT_FOUND, "Epic group not found"))?;
-    if detail.group.created_by != claims.user_id as i64 && claims.role != "root" {
+    if detail.group.created_by != claims.user_id && claims.role != "root" {
         return Err(err(StatusCode::FORBIDDEN, "Not owner"));
     }
     db::remove_epic_group_task(&engine.pool, id, task_id).await.map_err(internal)?;
@@ -70,7 +70,7 @@ pub async fn remove_epic_group_task(State(engine): State<AppState>, claims: Clai
 #[utoipa::path(post, path = "/api/epics/{id}/snapshot", responses((status = 204)), security(("bearer" = [])))]
 pub async fn snapshot_epic_group(State(engine): State<AppState>, claims: Claims, Path(id): Path<i64>) -> Result<StatusCode, ApiError> {
     let detail = db::get_epic_group_detail(&engine.pool, id).await.map_err(|_| err(StatusCode::NOT_FOUND, "Epic group not found"))?;
-    if detail.group.created_by != claims.user_id as i64 && claims.role != "root" {
+    if detail.group.created_by != claims.user_id && claims.role != "root" {
         return Err(err(StatusCode::FORBIDDEN, "Not owner"));
     }
     db::snapshot_epic_group(&engine.pool, id).await.map_err(internal)?;
